factor out convert text opts logic into file of its own
